# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Okta for Agents MVP API
  version: 1.0.0
servers:
  - url: http://localhost:8080/v1
paths:
  /health:
    get:
      summary: Service health check
      responses:
        "200":
          description: Healthy
  /metrics:
    get:
      summary: Route-level metrics snapshot
      responses:
        "200":
          description: Metrics snapshot
  /metrics/summary:
    get:
      summary: Global metrics summary
      responses:
        "200":
          description: Metrics summary
  /agents:
    get:
      summary: List agents for a tenant
      parameters:
        - in: query
          name: tenantId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Agents list
    post:
      summary: Create an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, name]
              properties:
                tenantId: { type: string }
                name: { type: string }
                environment: { type: string, enum: [dev, staging, prod] }
      responses:
        "201":
          description: Agent created
  /service-accounts:
    post:
      summary: Create service account machine principal
      responses:
        "201":
          description: Service account created
  /agents/{agentId}:
    delete:
      summary: Soft-delete an agent
      parameters:
        - in: path
          name: agentId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId]
              properties:
                tenantId: { type: string }
      responses:
        "200":
          description: Agent soft-deleted
  /policies:
    get:
      summary: List policies for a tenant
      parameters:
        - in: query
          name: tenantId
          required: true
          schema: { type: string }
        - in: query
          name: agentId
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Policies list
    post:
      summary: Create policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, agentId, connector, actions, environment, effect]
              properties:
                tenantId: { type: string }
                agentId: { type: string }
                connector: { type: string }
                actions:
                  type: array
                  items: { type: string }
                environment: { type: string, enum: [dev, staging, prod] }
                effect: { type: string, enum: [allow, deny, require_approval] }
      responses:
        "201":
          description: Policy created
  /policies/templates:
    get:
      summary: List reusable policy templates
      responses:
        "200":
          description: Template list
  /policies/simulate:
    post:
      summary: Simulate policy decision with optional proposed policy
      responses:
        "200":
          description: Simulation result
  /tokens/mint:
    post:
      summary: Mint short-lived token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, agentId]
              properties:
                tenantId: { type: string }
                agentId: { type: string }
      responses:
        "200":
          description: Token minted
  /tokens/revoke:
    post:
      summary: Revoke a JWT by token value
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, token]
              properties:
                tenantId: { type: string }
                token: { type: string }
                reason: { type: string }
      responses:
        "200":
          description: Token revoked
  /tenants/config:
    get:
      summary: Read tenant runtime config
      parameters:
        - in: query
          name: tenantId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tenant config
    post:
      summary: Upsert tenant runtime config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId]
              properties:
                tenantId: { type: string }
                tokenTtlSeconds: { type: integer }
      responses:
        "200":
          description: Tenant config upserted
  /tokens/keys:
    get:
      summary: List signing key metadata
      responses:
        "200":
          description: Key metadata list
    post:
      summary: Create a signing key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kid, secret]
              properties:
                kid: { type: string }
                secret: { type: string }
                activate: { type: boolean }
      responses:
        "201":
          description: Signing key created
  /tokens/keys/{kid}/activate:
    post:
      summary: Activate an existing signing key
      parameters:
        - in: path
          name: kid
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Signing key activated
  /authorize:
    post:
      summary: Evaluate policy decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, agentId, connector, action, environment]
              properties:
                tenantId: { type: string }
                agentId: { type: string }
                connector: { type: string }
                action: { type: string }
                environment: { type: string, enum: [dev, staging, prod] }
      responses:
        "200":
          description: Authorization decision
  /proxy/{connector}/{action}:
    post:
      summary: Execute policy-enforced connector action
      parameters:
        - in: path
          name: connector
          required: true
          schema: { type: string }
        - in: path
          name: action
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                environment: { type: string, enum: [dev, staging, prod] }
                approvalId: { type: string }
                payload: { type: object }
      responses:
        "200":
          description: Request allowed
        "202":
          description: Approval required
        "403":
          description: Request denied
  /approvals:
    get:
      summary: List approvals for a tenant
      parameters:
        - in: query
          name: tenantId
          required: true
          schema: { type: string }
        - in: query
          name: agentId
          required: false
          schema: { type: string }
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [pending, approved, rejected, consumed, expired]
      responses:
        "200":
          description: Approvals list
  /approvals/{approvalId}/decision:
    post:
      summary: Resolve approval
      parameters:
        - in: path
          name: approvalId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, approverId, decision]
              properties:
                tenantId: { type: string }
                approverId: { type: string }
                decision:
                  type: string
                  enum: [approved, rejected]
      responses:
        "200":
          description: Approval resolved
  /audit-events:
    get:
      summary: List audit events for a tenant
      parameters:
        - in: query
          name: tenantId
          required: true
          schema: { type: string }
        - in: query
          name: agentId
          required: false
          schema: { type: string }
        - in: query
          name: eventType
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Audit events list
  /policies/{policyId}:
    patch:
      summary: Update policy fields
      parameters:
        - in: path
          name: policyId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId]
              properties:
                tenantId: { type: string }
                connector: { type: string }
                actions:
                  type: array
                  items: { type: string }
                environment: { type: string, enum: [dev, staging, prod] }
                effect: { type: string, enum: [allow, deny, require_approval] }
      responses:
        "200":
          description: Policy updated
    delete:
      summary: Soft-delete policy
      parameters:
        - in: path
          name: policyId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId]
              properties:
                tenantId: { type: string }
      responses:
        "200":
          description: Policy soft-deleted
  /connectors/health:
    get:
      summary: Get connector health by tenant
      parameters:
        - in: query
          name: tenantId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Connector health details
  /connectors/github/credentials:
    post:
      summary: Set tenant GitHub credential
      responses:
        "200":
          description: Credential saved
  /connectors/slack/credentials:
    post:
      summary: Set tenant Slack credential
      responses:
        "200":
          description: Credential saved
  /connectors/jira/credentials:
    post:
      summary: Set tenant Jira credential
      responses:
        "200":
          description: Credential saved